Refactor: déplacer la classe Element dans son propre fichier
