+++ src/walle-lib/element.cpp
@@ -1,0 +1,7 @@
+#include "element.hpp"
+
+void Element::set_position(double x, double y)
+{
+    m_x = x;
+    m_y = y;
+}
